pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sdwan-orbit"
version = "0.1.0"
description = "Onboarding, Registration, Bootstrap & Integration Toolkit for Cisco SD-WAN"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "ORBIT Contributors"}
]
keywords = ["cisco", "sdwan", "vmanage", "onboarding", "automation", "ci-cd"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Networking",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "catalystwan>=0.33.10",
    "cisco-sdwan>=1.26",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "click>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.10",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
orbit = "sdwan_orbit.cli:cli"

[project.urls]
Homepage = "https://github.com/your-org/sdwan-orbit"
Documentation = "https://github.com/your-org/sdwan-orbit#readme"
Repository = "https://github.com/your-org/sdwan-orbit"
Issues = "https://github.com/your-org/sdwan-orbit/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
sdwan_orbit = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=sdwan_orbit --cov-report=term-missing"
